Add a template implementation of the photon+jet analysis
The class is dummy as it misses reading of data from input files and the
alpha extrapolation (if chosen to be implemented here). Computation of
chi2 needs to be checked.

The photon+jet analysis is added to the fit in the example program, but
at the moment it has no impact as the chi2 computed by it is always
zero.