Switch fit function in the example program to a standard one
Now that the photon+jet analysis is added, the fit can also determine
the absolute scale of the correction. Change the form of the fitted
correction to the standard two-parameter one.

To allow reusing corrections and to make the example program less
crowded, move definitions of standard corrections to a dedicated module.