Incorporate Z+jet measurement
Implemented by Viola in the same way as the photon+jet analysis. When
merging the corresponding branch, dropped nuisances in the Z+jet
analysis as they referred to the scale of the Z boson instead of muons
and electrons as would be done in reality. Also, will probably change
the implementation of nuisance parameters.