Refactor nuisances
Switch to using two classes to describe nuisances. The first one
(NuisanceDefinitions) is used to construct the list of relevant
nuisances at runtime. At this point they are not associated with any
concrete values. The second class (Nuisances) does not allow to register
new nuisances but provides values for already registered parameters.