Change binning in RecomputeBalanceData in multijet
Visualization histograms with recomputed mean balance in data in
MultijetCrawlingBins used binning in pt of the leading jet without the
L3Res correction applied. Now this correction is applied to bin edges so
that bins move along the x axis. The difference between this and the
histogram for simulation is more representative of the goodness of fit.