Do not exclude JER uncertainty
New JER-induced variations in L2Res are reasonable, so there is no need
to exclude this uncertainty anymore.

This uncertainty was first excluded in commit [1], although a number of
scipts were added for the first time after that commit, meaning that the
JER uncertainty was excluded in their initial versions.
[1] ded9ade9c887e8fc9f5505f1348ef9768f2dc742